Implement focus traversal (apart from https://github.com/derdilla/inline_tab_view/issues/1) and remove old addEntryDialoge implementation
